Move integration setup to coordinator `_async_setup` in Bring
