+++ homeassistant/components/bring/__init__.py
@@ -4,20 +4,13 @@ from __future__ import annotations
 
 import logging
 
-from bring_api import (
-    Bring,
-    BringAuthException,
-    BringParseException,
-    BringRequestException,
-)
+from bring_api import Bring
 
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_EMAIL, CONF_PASSWORD, Platform
 from homeassistant.core import HomeAssistant
-from homeassistant.exceptions import ConfigEntryAuthFailed, ConfigEntryNotReady
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 
-from .const import DOMAIN
 from .coordinator import BringDataUpdateCoordinator
 
 PLATFORMS: list[Platform] = [Platform.SENSOR, Platform.TODO]
@@ -30,30 +23,8 @@ type BringConfigEntry = ConfigEntry[BringDataUpdateCoordinator]
 async def async_setup_entry(hass: HomeAssistant, entry: BringConfigEntry) -> bool:
     """Set up Bring! from a config entry."""
 
-    email = entry.data[CONF_EMAIL]
-    password = entry.data[CONF_PASSWORD]
-
     session = async_get_clientsession(hass)
-    bring = Bring(session, email, password)
-
-    try:
-        await bring.login()
-    except BringRequestException as e:
-        raise ConfigEntryNotReady(
-            translation_domain=DOMAIN,
-            translation_key="setup_request_exception",
-        ) from e
-    except BringParseException as e:
-        raise ConfigEntryNotReady(
-            translation_domain=DOMAIN,
-            translation_key="setup_parse_exception",
-        ) from e
-    except BringAuthException as e:
-        raise ConfigEntryAuthFailed(
-            translation_domain=DOMAIN,
-            translation_key="setup_authentication_exception",
-            translation_placeholders={CONF_EMAIL: email},
-        ) from e
+    bring = Bring(session, entry.data[CONF_EMAIL], entry.data[CONF_PASSWORD])
 
     coordinator = BringDataUpdateCoordinator(hass, bring)
     await coordinator.async_config_entry_first_refresh()
